requirements.txt
opencv-python
Pillow
cryptography
tk
stegano
BitArray
scikit-image
numpy
customtkinter
